chore: moved padding to sha256 module
